Split the chat-messages module
