+++ packages/jupyter-chat/src/components/index.ts
@@ -3,12 +3,12 @@
  * Distributed under the terms of the Modified BSD License.
  */
 
+export * from './avatar';
 export * from './chat';
 export * from './chat-input';
-export * from './chat-messages';
 export * from './code-blocks';
 export * from './input';
 export * from './jl-theme-provider';
+export * from './messages';
 export * from './mui-extras';
 export * from './scroll-container';
-export * from './toolbar';
