+++ packages/jupyter-chat/src/components/messages/index.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Jupyter Development Team.
+ * Distributed under the terms of the Modified BSD License.
+ */
+
+export * from './footer';
+export * from './header';
+export * from './message';
+export * from './message-renderer';
+export * from './messages';
+export * from './navigation';
+export * from './toolbar';
+export * from './welcome';
+export * from './writers';
