+++ packages/jupyter-chat/src/components/messages/toolbar.tsx
